bookstore/requirements.txt
django
django_rest_framework
djangorestframework-simplejwt
django-cors-headers
django-debug-toolbar
django-filter